backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.0.0
reportlab>=4.0.0
python-dateutil>=2.8.0
Pillow>=10.0.0
PyMySQL>=1.1.0
mysql-connector-python>=8.0.33
cryptography>=41.0.0
openpyxl>=3.1.0
pandas>=2.0.0
waitress>=2.1.2
python-dotenv>=1.0.0
PyPDF2>=3.0.0
requests>=2.31.0
pytesseract>=0.3.10
pdf2image>=1.16.0
